requirements.txt
absl-py==2.4.0
accelerate==1.13.0
accelerated-scan==0.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-cors==0.8.1
aiosignal==1.4.0
aniso8601==10.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
asttokens==3.0.1
astunparse==1.6.3
attrs==26.1.0
awscli==1.44.65
babel==2.18.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==0.7.11
bokeh==3.9.0
boto3==1.42.75
botocore==1.42.75
braceexpand==0.1.7
brotli==1.2.0
cachetools==7.0.5
catalogue==2.0.10
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
colorful==0.5.8
comment_parser==1.2.4
confection==0.1.5
contourpy==1.3.3
cosmos-xenna==0.2.0
courlan==1.3.2
crossfit==0.0.9
cryptography==46.0.5
cuda-bindings==13.2.0
cuda-core==0.3.2
cuda-pathfinder==1.5.0
cuda-python==12.6.2.post1
cuda-toolkit==13.0.2
cudf-cu12==25.4.0
cudf-cu13==26.2.1
cugraph-cu12==25.4.0
cugraph-cu13==26.2.0
cuml-cu12==25.4.0
cuml-cu13==26.2.0
cupy-cuda12x==12.3.0
cupy-cuda13x==14.0.1
cuvs-cu12==25.4.0
cycler==0.12.1
cymem==2.0.13
dask==2025.2.0
dask-cuda==25.4.0
dask-cudf-cu12==25.4.0
dask-cudf-cu13==26.2.1
dask-mpi==2026.3.0
datasets==3.6.0
dateparser==1.3.0
decorator==5.2.1
dill==0.3.8
distlib==0.4.0
distributed==2025.2.0
distributed-ucxx-cu12==0.43.0
distributed-ucxx-cu13==0.48.0
distro==1.9.0
dm-tree==0.1.9
docutils==0.19
einops==0.8.2
executing==2.2.1
fasteners==0.20
fastrlock==0.8.3
fasttext==0.9.3
FastWARC==0.15.2
fiddle==0.3.0
filelock==3.25.2
Flask==3.1.3
Flask-RESTful==0.3.10
frozenlist==1.8.0
fsspec==2025.3.0
ftfy==6.1.1
future==1.0.0
gast==0.7.0
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-auth==2.49.1
googleapis-common-protos==1.73.0
graphviz==0.21
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
in-place==0.5.0
inflect==7.5.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jusText==3.0.1
langcodes==3.5.1
libcudf-cu12==25.4.0
libcudf-cu13==26.2.1
libcugraph-cu12==25.4.1
libcugraph-cu13==26.2.0
libcuml-cu12==25.4.0
libcuml-cu13==26.2.0
libcuvs-cu12==25.4.0
libkvikio-cu12==25.4.0
libkvikio-cu13==26.2.0
libraft-cu12==25.4.0
libraft-cu13==26.2.0
librmm-cu12==25.4.0
librmm-cu13==26.2.0
libucx-cu12==1.18.1
libucx-cu13==1.19.0
libucxx-cu12==0.43.0
libucxx-cu13==0.48.0
lightning-utilities==0.15.3
llvmlite==0.43.0
locket==1.0.0
loguru==0.7.3
lxml==6.0.2
lxml_html_clean==0.4.4
lz4==4.4.5
makefun==1.16.0
Markdown==3.10.2
markdown-it-py==4.0.0
markdown2==2.5.5
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mecab-python3==1.0.12
more-itertools==10.8.0
mpi4py==4.1.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.16
murmurhash==1.0.15
mwparserfromhell==0.6.5
narwhals==2.18.1
nemo-curator @ file:///tmp/nemo-curator
networkx==3.6.1
ninja==1.13.0
numba==0.60.0
numba-cuda==0.4.0
numpy==1.26.4
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.9.2.10
nvidia-cuda-cccl==13.2.27
nvidia-cuda-crt==13.2.51
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvcc==13.0.88
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu13==0.8.0
nvidia-dali-cuda120==2.0.0
nvidia-libnvcomp-cu12==5.1.0.21
nvidia-libnvcomp-cu13==5.1.0.21
nvidia-ml-py==12.575.51
nvidia-nccl-cu13==2.28.9
nvidia-nvcomp-cu12==4.2.0.11
nvidia-nvimgcodec-cu12==0.7.0.11
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvjpeg-cu12==12.4.0.76
nvidia-nvjpeg2k-cu12==0.9.1.47
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtiff-cu12==0.6.0.78
nvidia-nvtx==13.0.85
nvidia-nvvm==13.2.51
nvtx==0.2.15
obstore==0.9.2
omegaconf==2.3.0
openai==2.29.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.40.0
opentelemetry-exporter-prometheus==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging @ file:///home/task_176104874243446/conda-bld/packaging_1761049080023/work
pandas==2.2.3
pangu==4.0.6.1
parso==0.8.6
partd==1.4.2
peft==0.18.1
pexpect==4.9.0
phonenumbers==8.13.55
pillow==12.1.1
platformdirs==4.9.4
portalocker==3.2.0
portpicker==1.6.0
preshed==3.0.13
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351
probableparsing==0.0.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
PuLP==3.3.0
pure_eval==0.2.3
py-spy==0.4.1
pyarrow==19.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybind11==3.0.2
pycld2==0.42
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pylibcudf-cu12==25.4.0
pylibcudf-cu13==26.2.1
pylibcugraph-cu12==25.4.1
pylibcugraph-cu13==26.2.0
pylibraft-cu12==25.4.0
pylibraft-cu13==26.2.0
pynvjitlink-cu12==0.7.0
pynvml==12.0.0
pyparsing==3.3.2
PySocks==1.7.1
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-discovery==1.2.0
python-magic==0.4.24
pytorch-lightning==2.6.1
pytz==2026.1.post1
PyYAML==6.0.3
raft-dask-cu12==25.4.0
raft-dask-cu13==26.2.0
rapids-dask-dependency==25.4.0
rapids-logger==0.1.19
ray==2.54.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-file==3.0.1
Resiliparse==0.15.2
rich==14.3.3
rmm-cu12==25.4.0
rmm-cu13==26.2.0
rpds-py==0.30.0
rsa==4.7.2
ruamel.yaml==0.19.1
s3transfer==0.16.0
sacrebleu==2.6.0
sacremoses==0.1.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
sentence-transformers==5.3.0
sentencepiece==0.2.1
sentry-sdk==2.56.0
setuptools==80.10.2
setuptools-rust==1.12.0
setuptools-scm==9.2.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
stack-data==0.6.3
sympy==1.14.0
tabulate==0.10.0
tblib==3.2.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow-metadata==1.17.3
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.6.0
timm==1.0.26
tld==0.13.2
tldextract==5.3.1
tokenizers==0.22.2
toolz==1.1.0
torch==2.11.0+cu130
torchmetrics==1.9.0
torchvision==0.26.0+cu130
tornado==6.5.5
tqdm==4.67.3
trafilatura==2.0.0
traitlets==5.14.3
transformers==5.3.0
treelite==4.4.1
triton==3.6.0
typeguard==4.5.1
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing-utils==0.1.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ucx-py-cu12==0.43.0
ucxx-cu12==0.43.0
ucxx-cu13==0.48.0
unidic-lite==1.0.8
urllib3==2.6.3
usaddress==0.5.10
uv==0.11.1
virtualenv==21.2.0
warcio==1.7.4
wasabi==1.1.3
wcwidth==0.6.0
weasel==0.4.3
webdataset==1.0.2
Werkzeug==3.1.7
wheel==0.46.3
wheel-stub==0.5.0
wrapt==2.1.2
xxhash==3.6.0
xyzservices==2025.11.0
yarl==1.23.0
zict==3.0.0
zipp==3.23.0
zstandard==0.18.0
